requirements.txt
telethon>=1.34.0
python-dotenv>=1.0.0
PyYAML>=6.0
